refactor: migrate GitHub test suite to GitHubClientTests (#1909)
Moves the six transport/rate-limit/URL/counter test files from
Tests/RunBotCoreTests/GitHub/ into the new Tests/GitHubClientTests target and
switches them to `@testable import GitHubClient`. Adds keychain-free test
doubles local to the new target: MockTokenStore (Mutex-backed TokenStore) and
SpyRateLimitActor (extracted from RunBotCoreTests/TestDoubles.swift, which is
trimmed accordingly). Rewrites GitHubTokenCacheTests to exercise the injectable
GitHubClient.TokenCache directly via MockTokenStore, replacing the old
env/Keychain-based free-function tests.
