+++ Tests/GitHubClientTests/APICallCounter+TestSeam.swift
@@ -1,11 +1,11 @@
 // APICallCounter+TestSeam.swift
-// RunBotCoreTests
+// GitHubClientTests
 //
 // Test-only extensions on APICallCounter for seeding and resetting state
 // without real time travel. Compiled only in the test target.
 import Foundation
 
-@testable import RunBotCore
+@testable import GitHubClient
 
 extension APICallCounter {
   /// Seeds the rolling buffer with pre-built `ContinuousClock.Instant` values.
